refactor, dom 관련 작업 함수 별도로 분리
